Refactor the openstack provider code
This PR refactors/moves the provider code for better readability.
