+++ pkg/util/errors/errors.go
@@ -17,10 +17,27 @@ limitations under the License.
 package errors
 
 import (
+	"errors"
 	"net/http"
 
 	"github.com/gophercloud/gophercloud"
 )
+
+// ErrNotFound is used to inform that the object is missing
+var ErrNotFound = errors.New("failed to find object")
+
+// ErrMultipleResults is used when we unexpectedly get back multiple results
+var ErrMultipleResults = errors.New("multiple results where only one expected")
+
+// ErrNoAddressFound is used when we cannot find an ip address for the host
+var ErrNoAddressFound = errors.New("no address found for host")
+
+// ErrIPv6SupportDisabled is used when one tries to use IPv6 Addresses when
+// IPv6 support is disabled by config
+var ErrIPv6SupportDisabled = errors.New("IPv6 support is disabled")
+
+//ErrNoRouterID is used when router-id is not set
+var ErrNoRouterID = errors.New("router-id not set in cloud provider config")
 
 func IsNotFound(err error) bool {
 	if _, ok := err.(gophercloud.ErrDefault404); ok {
